app/requirements.txt
numpy==2.1.*
pandas==2.2.*
yfinance==0.2.*
fastapi==0.115.*
pydantic-settings==2.5.*
uvicorn[standard]
redis[async]